factor out makeBufferWithContents
